move account/txConfirmLogic -> tx/txConfirmLogic
